bundle/direct/dstate: make VersionID the recorded version, int all the way
Under recording the version is the counter that matters, so the state carries it in
an exported VersionID instead of borrowing Data.Serial. Open anchors it to the
deployment last version, startVersion advances it in place when it creates one, and
readers take it from there - so nothing has to recompute or pass it along.

The plan still stamps its serial from it, and validatePlan compares against it, so
stale plans are rejected exactly as before. Both directions are covered in one
branch: behind the recorded version gets the version-worded error, anything else
falls through to the serial comparison, whose numbers agree for both backends.

The version is an int end to end now, through the DMS client and the URL helper, so
there are no int64 conversions left at the comparison sites.

logDeploymentVersion reads the deployment and version off the state rather than
taking them as arguments.

+++ libs/dms/metadata_test.go
@@ -10,7 +10,7 @@ import (
 func TestNextVersion(t *testing.T) {
 	for _, tt := range []struct {
 		last string
-		want int64
+		want int
 	}{
 		{"", 1},
 		{"4", 5},
